feature: left paddle movable and stays in bounds.
created necessary constants and logic to move the left paddle and keep it in bounds. Also have the ball and other paddle in place.

+++ src/main.cpp
@@ -3,8 +3,8 @@
 
 int main() {
 
-    constexpr size_t framesPerSecond{60};
-    constexpr size_t msPerFrame{1000 / framesPerSecond};
+    constexpr Uint32 framesPerSecond{60};
+    constexpr Uint32 msPerFrame{1000 / framesPerSecond};
     constexpr size_t screenWidth{1024};
     constexpr size_t screenHeight{768};
